+++ internal/daemon/exec_bpf.go
@@ -70,7 +70,7 @@ func (e *execBPF) Run(ctx context.Context) {
 	}()
 
 	go e.reader.Run(ctx)
-	pcCache, _ := ProcessCtx()
+	pcCache, pcEnr := ProcessCtx()
 	for {
 		select {
 		case <-ctx.Done():
@@ -81,6 +81,9 @@ func (e *execBPF) Run(ctx context.Context) {
 			}
 			e.count.Add(1)
 			populateProcessCtxFromExec(pcCache, ev)
+			if ev.PID != 0 {
+				pcEnr.Enqueue(processctxRequestFromExec(ev))
+			}
 			for _, f := range checks.EvaluateExec(ev.UID, ev.PID, ev.Comm, ev.Filename, ev.ParentComm) {
 				attachProcessCtxToExecFinding(pcCache, &f, ev)
 				select {